requirements.txt
Flask==1.1.1
numpy==1.15.4
requests==2.20.1
Pillow==5.3.0